+++ api/.env
@@ -1,0 +1,1 @@
+SECRET_KEY="secret"
