+++ api/app.js
@@ -3,6 +3,7 @@ const cors = require('cors');
 const app = express();
 const zetter = require('./data/tweets.js');
 const zetter_routes = require('./routes/zetter.js');
+require('dotenv').config();
 
 app.listen(5000, () => {
     console.log('server is listening on port 5000');
